[PROT-386] Improvemnt in Time strucutres TODO:// Check Constraints TODO:// Refactor LibMediaItem to fill LibFileItem TODO:// Implemments test for Building SchEventConfiguration SchClockConfiguration and grid TODO:// Log upload pipe TODO:// Finish merging TODO:// Fix unit test TODO:// Write unit tests

+++ scheduler/src/main/java/io/protone/scheduler/domain/SchTimeParams.java
@@ -1,7 +1,5 @@
 package io.protone.scheduler.domain;
 
-import io.protone.core.domain.AbstractAuditingEntity;
-
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.io.Serializable;
@@ -22,7 +20,7 @@ public class SchTimeParams implements Serializable {
     private ZonedDateTime endTime;
 
     @Column(name = "relative_delay")
-    private Integer relativeDelay;
+    private Long relativeDelay;
 
 
     public ZonedDateTime getStartTime() {
@@ -51,15 +49,15 @@ public class SchTimeParams implements Serializable {
         return this;
     }
 
-    public Integer getRelativeDelay() {
+    public Long getRelativeDelay() {
         return relativeDelay;
     }
 
-    public void setRelativeDelay(Integer relativeDelay) {
+    public void setRelativeDelay(Long relativeDelay) {
         this.relativeDelay = relativeDelay;
     }
 
-    public SchTimeParams relativeDelay(Integer relativeDelay) {
+    public SchTimeParams relativeDelay(Long relativeDelay) {
         this.relativeDelay = relativeDelay;
         return this;
     }
